Move autostop to separate file/object
